refactor: split sm_v003 into state_machine and snapshot_store
# Summary

`sm_v003` is replaced by two modules named for what they contain: the
`state_machine` that applies raft entries and serves reads, and the
`snapshot_store` that builds, ships and loads snapshot files. Its V002-to-V004
entry adapter moves to `sled_compat`.

# Details

The name `sm_v003` was wrong twice over. It held the current V004 code, not
V003, and it held two unrelated subsystems that never call each other: the
state machine hands the snapshot store a `DB` and takes one back, and that is
the whole interface between them. Nineteen files under one version-stamped
directory hid that.

`adapter` belongs to neither. It converts V002 sled entries into V004 rotbl
pairs and is only reachable from the upgrade path, so it joins the rest of the
legacy layout in `sled_compat`.

Type names still carry their `V003`/`V004` suffixes; renaming them is a
separate change so this one stays a pure move.

Public paths change accordingly:

    sm_v003::SMV003  -> state_machine::SMV003
    sm_v003::<rest>  -> snapshot_store::<rest>
