+++ crates/server/raft-store/src/immutable_compactor/mod.rs
@@ -21,7 +21,7 @@ use log::info;
 use crate::leveled_store::leveled_map::CompactorPermit;
 use crate::leveled_store::leveled_map::LeveledMap;
 use crate::leveled_store::leveled_map::WriterPermit;
-use crate::sm_v003::SMV003;
+use crate::state_machine::SMV003;
 
 #[cfg(test)]
 mod immutable_compactor_test;
