+++ crates/server/raft-store/src/ondisk/upgrade_to_v004.rs
@@ -35,8 +35,8 @@ use crate::raft_log_v004::importer;
 use crate::sled_compat::LogMetaKey;
 use crate::sled_compat::key_spaces::LogMeta;
 use crate::sled_compat::key_spaces::RaftStoreEntry;
-use crate::sm_v003::SnapshotStoreV003;
-use crate::sm_v003::SnapshotStoreV004;
+use crate::snapshot_store::SnapshotStoreV003;
+use crate::snapshot_store::SnapshotStoreV004;
 
 impl OnDisk {
     /// Upgrade the on-disk data form [`DataVersion::V003`] to [`DataVersion::V004`].
