+++ crates/server/raft-store/src/sled_compat/adapter.rs
