+++ crates/server/raft-store/src/snapshot_store/open_snapshot.rs
