+++ crates/server/raft-store/src/snapshot_store/snapshot_loader.rs
@@ -22,10 +22,10 @@ use log::info;
 use log::warn;
 use openraft::SnapshotId;
 
-use crate::sm_v003::MetaSnapshotId;
-use crate::sm_v003::SnapshotStoreError;
-use crate::sm_v003::open_snapshot::OpenSnapshot;
 use crate::snapshot_config::SnapshotConfig;
+use crate::snapshot_store::MetaSnapshotId;
+use crate::snapshot_store::SnapshotStoreError;
+use crate::snapshot_store::open_snapshot::OpenSnapshot;
 
 /// Load snapshot from persisted storage.
 ///
