+++ crates/server/raft-store/src/snapshot_store/snapshot_store_error.rs
