+++ crates/server/raft-store/src/snapshot_store/write_entry.rs
