+++ crates/server/raft-store/src/state_machine/kv_api.rs
@@ -32,7 +32,7 @@ use state_machine_api::UserKey;
 
 use crate::leveled_store::state_machine::read_view::StateMachineReadView;
 use crate::leveled_store::util::add_cooperative_yielding;
-use crate::sm_v003::SMV003;
+use crate::state_machine::SMV003;
 use crate::testing::since_epoch_millis;
 use crate::utils::prefix_right_bound;
 use crate::utils::seq_marked_to_seqv;
