+++ crates/server/service/src/meta_service/raft_service_impl.rs
@@ -22,12 +22,12 @@ use databend_base::counter::Counter;
 use databend_meta_client::MetaGrpcReadReq;
 use databend_meta_raft_store::data_version::DATA_VERSION;
 use databend_meta_raft_store::leveled_store::persisted_codec::PersistedCodec;
-use databend_meta_raft_store::sm_v003::MetaSnapshotId;
-use databend_meta_raft_store::sm_v003::WriterV003;
-use databend_meta_raft_store::sm_v003::open_snapshot::OpenSnapshot;
-use databend_meta_raft_store::sm_v003::received::Received;
-use databend_meta_raft_store::sm_v003::write_entry::WriteEntry;
 use databend_meta_raft_store::snapshot_config::SnapshotConfig;
+use databend_meta_raft_store::snapshot_store::MetaSnapshotId;
+use databend_meta_raft_store::snapshot_store::WriterV003;
+use databend_meta_raft_store::snapshot_store::open_snapshot::OpenSnapshot;
+use databend_meta_raft_store::snapshot_store::received::Received;
+use databend_meta_raft_store::snapshot_store::write_entry::WriteEntry;
 use databend_meta_runtime_api::SpawnApi;
 use databend_meta_snapshot_db::DB;
 use databend_meta_snapshot_db::Snapshot;
