+++ crates/server/service/src/store/meta_raft_state_machine/mod.rs
@@ -24,10 +24,10 @@ use std::time::Duration;
 use databend_base::counter::Counter;
 use databend_meta_raft_store::config::RaftConfig;
 use databend_meta_raft_store::immutable_compactor::InMemoryCompactor;
-use databend_meta_raft_store::sm_v003::MetaSnapshotId;
-use databend_meta_raft_store::sm_v003::SMV003;
-use databend_meta_raft_store::sm_v003::SnapshotStoreV004;
-use databend_meta_raft_store::sm_v003::WriteEntry;
+use databend_meta_raft_store::snapshot_store::MetaSnapshotId;
+use databend_meta_raft_store::snapshot_store::SnapshotStoreV004;
+use databend_meta_raft_store::snapshot_store::WriteEntry;
+use databend_meta_raft_store::state_machine::SMV003;
 use databend_meta_runtime_api::SpawnApi;
 use databend_meta_snapshot_db::DB;
 use databend_meta_snapshot_db::Snapshot;
