+++ crates/server/service/src/store/store.rs
@@ -30,9 +30,9 @@ use databend_meta_raft_store::raft_log_v004::Cw;
 use databend_meta_raft_store::raft_log_v004::RaftLogV004;
 use databend_meta_raft_store::raft_log_v004::util;
 use databend_meta_raft_store::sled_compat::key_spaces::RaftStoreEntry;
-use databend_meta_raft_store::sm_v003::MetaSnapshotId;
-use databend_meta_raft_store::sm_v003::SMV003;
-use databend_meta_raft_store::sm_v003::SnapshotStoreV004;
+use databend_meta_raft_store::snapshot_store::MetaSnapshotId;
+use databend_meta_raft_store::snapshot_store::SnapshotStoreV004;
+use databend_meta_raft_store::state_machine::SMV003;
 use databend_meta_runtime_api::SpawnApi;
 use databend_meta_snapshot_db::DB;
 use databend_meta_snapshot_db::DBStat;
